Renamed User_Interface folder for consistency.
