+++ Source/Pawnmorphs/Esoteria/Pawnmorph.csproj
@@ -571,11 +571,11 @@
     <Compile Include="Jobs\Driver_UseGenome.cs" />
     <Compile Include="Jobs\Driver_UseMutationScanner.cs" />
     <Compile Include="Jobs\Giver_FoodOptimized.cs" />
-    <Compile Include="User Interface\Genebank\GeneRowItem.cs" />
-    <Compile Include="User Interface\Genebank\GenebankTab.cs" />
-    <Compile Include="User Interface\Genebank\Tabs\AnimalsTab.cs" />
-    <Compile Include="User Interface\Genebank\Tabs\MutationsTab.cs" />
-    <Compile Include="User Interface\MainTabWindow_ChamberDatabase.cs" />
+    <Compile Include="UserInterface\Genebank\GeneRowItem.cs" />
+    <Compile Include="UserInterface\Genebank\GenebankTab.cs" />
+    <Compile Include="UserInterface\Genebank\Tabs\AnimalsTab.cs" />
+    <Compile Include="UserInterface\Genebank\Tabs\MutationsTab.cs" />
+    <Compile Include="UserInterface\MainTabWindow_ChamberDatabase.cs" />
     <Compile Include="MergedPawnUtilities.cs" />
     <Compile Include="ModInitializationException.cs" />
     <Compile Include="InstinctUtilities.cs" />
@@ -871,24 +871,24 @@
     <Compile Include="ToolVerbUtilities.cs" />
     <Compile Include="TransformerUtility.cs" />
     <Compile Include="TransformPawn.cs" />
-    <Compile Include="User Interface\AddedMutations.cs" />
-    <Compile Include="User Interface\ButtonTexturesPM.cs" />
-    <Compile Include="User Interface\Dialog_PartPicker.cs" />
-    <Compile Include="User Interface\Dialog_Popup.cs" />
-    <Compile Include="User Interface\FilterListBox.cs" />
-    <Compile Include="User Interface\MutationData.cs" />
-    <Compile Include="User Interface\Preview\DrawOptions.cs" />
-    <Compile Include="User Interface\Preview\HumanlikePreview.cs" />
-    <Compile Include="User Interface\Preview\Preview.cs" />
-    <Compile Include="User Interface\Preview\PawnPreview.cs" />
-    <Compile Include="User Interface\Preview\ThingPreview.cs" />
-    <Compile Include="User Interface\Settings\Dialog_OptionalPatches.cs" />
-    <Compile Include="User Interface\Settings\Dialog_RaceReplacements.cs" />
-    <Compile Include="User Interface\Settings\Dialog_VisibleRaceSelection.cs" />
-    <Compile Include="User Interface\TableBox\ITableRow.cs" />
-    <Compile Include="User Interface\TableBox\Table.cs" />
-    <Compile Include="User Interface\TableBox\TableColumn.cs" />
-    <Compile Include="User Interface\Window_Genebank.cs" />
+    <Compile Include="UserInterface\AddedMutations.cs" />
+    <Compile Include="UserInterface\ButtonTexturesPM.cs" />
+    <Compile Include="UserInterface\Dialog_PartPicker.cs" />
+    <Compile Include="UserInterface\Dialog_Popup.cs" />
+    <Compile Include="UserInterface\FilterListBox.cs" />
+    <Compile Include="UserInterface\MutationData.cs" />
+    <Compile Include="UserInterface\Preview\DrawOptions.cs" />
+    <Compile Include="UserInterface\Preview\HumanlikePreview.cs" />
+    <Compile Include="UserInterface\Preview\Preview.cs" />
+    <Compile Include="UserInterface\Preview\PawnPreview.cs" />
+    <Compile Include="UserInterface\Preview\ThingPreview.cs" />
+    <Compile Include="UserInterface\Settings\Dialog_OptionalPatches.cs" />
+    <Compile Include="UserInterface\Settings\Dialog_RaceReplacements.cs" />
+    <Compile Include="UserInterface\Settings\Dialog_VisibleRaceSelection.cs" />
+    <Compile Include="UserInterface\TableBox\ITableRow.cs" />
+    <Compile Include="UserInterface\TableBox\Table.cs" />
+    <Compile Include="UserInterface\TableBox\TableColumn.cs" />
+    <Compile Include="UserInterface\Window_Genebank.cs" />
     <Compile Include="Utilities\Cached.cs" />
     <Compile Include="Utilities\Collections\FilteredList.cs" />
     <Compile Include="Utilities\CompCacher.cs" />
