+++ Source/Pawnmorphs/Esoteria/UserInterface/Preview/HumanlikePreview.cs
@@ -10,7 +10,7 @@ using System.Threading.Tasks;
 using UnityEngine;
 using Verse;
 
-namespace Pawnmorph.User_Interface.Preview
+namespace Pawnmorph.UserInterface.Preview
 {
     internal class HumanlikePreview : ThingPreview
     {
