+++ Source/Pawnmorphs/Esoteria/UserInterface/TableBox/ITableRow.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Pawnmorph.User_Interface.TableBox
+namespace Pawnmorph.UserInterface.TableBox
 {
     internal interface ITableRow
     {
